requirements.txt
pandas
numpy
requests
scikit-learn
scipy
python-dotenv
matplotlib
kalshi_python_sync
ngboost
